chore(interface): Restructure interface
